Send notifications as background tasks. Closes #598
