+++ docs/running/index.rst
@@ -187,3 +187,12 @@ We describe using Dokku below, and assume that you have already have `Dokku inst
    * Under **Indigo API** click Countries, then click Add Country in the top right corner
    * Choose a country and primary language from the dropdown lists
    * Click Save
+
+Background Tasks
+----------------
+
+Indigo can optionally do some operations in the background. It requires a worker or
+cron job to run the ``django-background-tasks`` task queue. See `django-background-tasks <https://django-background-tasks.readthedocs.io/en/latest/>`
+for more details on running background tasks.
+
+To enable background tasks, set ``INDIGO.NOTIFICATION_EMAILS_BACKGROUND`` to True.
