+++ setup.py
@@ -48,6 +48,7 @@ setup(
         'dj-database-url>=0.3.0',
         'django-activity-stream>=0.7.0',
         'django-allauth>=0.36.0',
+        'django-background-tasks>=1.2.0',
         'django-compressor>=2.2',
         'django-cors-headers>1.3.1',
         'django-countries-plus==1.1',  # 1.2 breaks migrations
